3. Интерфейс Input [#181092 #364339]

+++ src/main/java/ru/job4j/poly/Transport.java
@@ -5,5 +5,5 @@ public interface Transport {
 
     void passengers(int numbers);
 
-    int refuel(int quantity);
+    float refuel(float quantity);
 }
